+++ integration-tests/solidity-abi/solidity-calls-flipper/lib.rs
@@ -19,7 +19,7 @@ pub mod flipper {
         /// Creates a new flipper smart contract initialized with the given value.
         #[ink(constructor)]
         pub fn new(init_value: bool) -> Self {
-            assert!(init_value == false);
+            assert!(!init_value);
             Self { value: false }
         }
